test(datetime): migrate minmax
